+++ service-implementation/src/main/java/com/liaison/mailbox/service/core/FileStageReplicationService.java
@@ -10,24 +10,27 @@
 
 package com.liaison.mailbox.service.core;
 
+import com.liaison.commons.jaxb.JAXBUtility;
 import com.liaison.commons.logging.LogTags;
-import com.liaison.commons.messagebus.client.exceptions.ClientUnavailableException;
 import com.liaison.mailbox.MailBoxConstants;
 import com.liaison.mailbox.enums.EntityStatus;
+import com.liaison.mailbox.enums.FailoverMessageType;
 import com.liaison.mailbox.rtdm.dao.StagedFileDAO;
 import com.liaison.mailbox.rtdm.dao.StagedFileDAOBase;
 import com.liaison.mailbox.rtdm.model.StagedFile;
+import com.liaison.mailbox.service.directory.DirectoryService;
 import com.liaison.mailbox.service.exception.MailBoxServicesException;
 import com.liaison.mailbox.service.queue.sender.FileStageReplicationSendQueue;
 import com.liaison.mailbox.service.storage.util.StorageUtilities;
 import com.liaison.mailbox.service.util.DirectoryCreationUtil;
+import com.liaison.mailbox.service.util.FileWriterUtil;
 import com.liaison.mailbox.service.util.MailBoxUtil;
-
+import com.liaison.mailbox.service.util.ShellScriptEngineUtil;
+import com.liaison.usermanagement.service.dto.DirectoryMessageDTO;
 import org.apache.commons.io.IOUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.apache.logging.log4j.ThreadContext;
-import org.codehaus.jettison.json.JSONException;
 import org.codehaus.jettison.json.JSONObject;
 
 import java.io.File;
@@ -38,15 +41,19 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 
+import static com.liaison.mailbox.MailBoxConstants.DIRECTORY_STAGE_REPLICATION_RETRY_DELAY;
+import static com.liaison.mailbox.MailBoxConstants.DIRECTORY_STAGE_REPLICATION_RETRY_MAX_COUNT;
+import static com.liaison.mailbox.MailBoxConstants.FAILOVER_MSG_TYPE;
 import static com.liaison.mailbox.MailBoxConstants.FILE_STAGE_REPLICATION_RETRY_DELAY;
 import static com.liaison.mailbox.MailBoxConstants.FILE_STAGE_REPLICATION_RETRY_MAX_COUNT;
 import static com.liaison.mailbox.MailBoxConstants.GLOBAL_PROCESS_ID;
-import static com.liaison.mailbox.MailBoxConstants.RETRY_COUNT;
-import static com.liaison.mailbox.MailBoxConstants.URI;
-import static com.liaison.mailbox.MailBoxConstants.KEY_PROCESSOR_ID;
 import static com.liaison.mailbox.MailBoxConstants.KEY_FILE_NAME;
 import static com.liaison.mailbox.MailBoxConstants.KEY_FILE_PATH;
 import static com.liaison.mailbox.MailBoxConstants.KEY_OVERWRITE;
+import static com.liaison.mailbox.MailBoxConstants.MESSAGE;
+import static com.liaison.mailbox.MailBoxConstants.RETRY_COUNT;
+import static com.liaison.mailbox.MailBoxConstants.TRIGGER_FILE;
+import static com.liaison.mailbox.MailBoxConstants.URI;
 
 
 /**
@@ -60,6 +67,9 @@ public class FileStageReplicationService implements Runnable {
 
     private static final Long DELAY = MailBoxUtil.getEnvironmentProperties().getLong(FILE_STAGE_REPLICATION_RETRY_DELAY, 10000);
     private static final Long MAX_RETRY_COUNT = MailBoxUtil.getEnvironmentProperties().getLong(FILE_STAGE_REPLICATION_RETRY_MAX_COUNT, 10);
+
+    private static final Long DIR_DELAY = MailBoxUtil.getEnvironmentProperties().getLong(DIRECTORY_STAGE_REPLICATION_RETRY_DELAY, 10000);
+    private static final Long DIR_MAX_RETRY_COUNT = MailBoxUtil.getEnvironmentProperties().getLong(DIRECTORY_STAGE_REPLICATION_RETRY_MAX_COUNT, 10);
 
     private String message;
 
@@ -81,7 +91,22 @@ public class FileStageReplicationService implements Runnable {
 
     @Override
     public void run() {
-        this.stage(getMessage());
+        this.doProcess(getMessage());
+    }
+
+    public void doProcess(String requestMessage) {
+
+        try {
+            JSONObject requestObj = new JSONObject(requestMessage);
+            String type = requestObj.getString(FAILOVER_MSG_TYPE);
+            if (FailoverMessageType.FILE.name().equalsIgnoreCase(type)) {
+                stage(requestMessage, requestObj);
+            } else {
+                stageDirectory(requestMessage, requestObj);
+            }
+        } catch (Throwable e) {
+            LOGGER.error(e.getMessage(), e);
+        }
     }
 
     /**
@@ -89,16 +114,17 @@ public class FileStageReplicationService implements Runnable {
      *
      * @param requestMessage message string from the queue
      */
-    public void stage(String requestMessage) {
+    private void stage(String requestMessage, JSONObject requestObj) {
 
         try {
 
-            JSONObject requestObj = new JSONObject(requestMessage);
             String fs2uri = requestObj.getString(URI);
             String globalProcessId = requestObj.getString(GLOBAL_PROCESS_ID);
             String fileName = requestObj.getString(KEY_FILE_NAME);
             String processorPayloadLocation = requestObj.getString(KEY_FILE_PATH);
             String isOverwrite = requestObj.getString(KEY_OVERWRITE);
+            boolean isTriggerFile = requestObj.getBoolean(TRIGGER_FILE);
+
             InputStream payload = null;
 
             int retry = (int) requestObj.get(RETRY_COUNT);
@@ -120,6 +146,11 @@ public class FileStageReplicationService implements Runnable {
 
                 if (EntityStatus.INACTIVE.value().equalsIgnoreCase(stagedFile.getStagedFileStatus())) {
                     LOGGER.warn("Staged file is inactive status and there is no need to stage it - {}", globalProcessId);
+                    return;
+                }
+
+                if (isTriggerFile) {
+                    FileWriterUtil.writeTriggerFile(processorPayloadLocation, fileName, fs2uri);
                     return;
                 }
 
@@ -182,6 +213,49 @@ public class FileStageReplicationService implements Runnable {
                 outputStream.close();
             }
         }
+    }
+
+    public void postDirectoryMessageToQueue(DirectoryMessageDTO message, int retryCount) {
+
+        try {
+            JSONObject directoryObject = new JSONObject();
+            directoryObject.put(FAILOVER_MSG_TYPE, FailoverMessageType.DIRECTORY);
+            directoryObject.put(MESSAGE, JAXBUtility.marshalToJSON(message));
+            directoryObject.put(RETRY_COUNT, retryCount);
+            FileStageReplicationSendQueue.getInstance().sendMessage(directoryObject.toString(), DIR_DELAY);
+        } catch (Exception e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    private void stageDirectory(String requestMessage, JSONObject requestObj) {
+
+        try {
+
+            int retry = 0;
+            if (requestObj.has(RETRY_COUNT)) {
+
+                retry = (int) requestObj.get(RETRY_COUNT);
+                if (retry > DIR_MAX_RETRY_COUNT) {
+                    LOGGER.warn("Reached maximum retry and dropping this message - {}", requestMessage);
+                    return;
+                }
+            }
+
+            //stageDirectory creation logic
+            DirectoryMessageDTO dto = JAXBUtility.unmarshalFromJSON(requestObj.getString(MESSAGE), DirectoryMessageDTO.class);
+            if (ShellScriptEngineUtil.validateUser(dto.getUserName())) {
+                new DirectoryService("").invokeScriptToCreateFolderAndAssignPermission(dto.getGatewayType(), dto.getUserName().toLowerCase());
+            } else {
+                LOGGER.warn("User {} is not replicated yet", dto.getUserName());
+                //post back to queue if AD is not synced yet
+                postDirectoryMessageToQueue(dto, ++retry);
+            }
+
+        } catch (Exception e) {
+            throw new RuntimeException(e);
+        }
+
     }
 
 }