+++ service-implementation/src/main/java/com/liaison/mailbox/service/dto/dropbox/StagedFileDTO.java
@@ -216,6 +216,7 @@ public class StagedFileDTO {
 		this.setId(file.getPguid());
         this.setClusterType(file.getClusterType());
         this.setParentGlobalProcessId(file.getParentGlobalProcessId());
+        this.setProcessDc(file.getProcessDc());
     }
 
 }